+++ noodles-bam/src/bai/async.rs
@@ -1,12 +1,12 @@
 //! Async BAI.
 
 pub mod io;
-mod writer;
 
 #[deprecated(since = "0.68.0", note = "Use `bai::r#async::io::Reader` instead.")]
 pub use self::io::Reader;
 
-pub use self::writer::Writer;
+#[deprecated(since = "0.68.0", note = "Use `bai::r#async::io::Writer` instead.")]
+pub use self::io::Writer;
 
 use std::path::Path;